finish request staff

+++ app/controllers/HotelController.php
@@ -62,14 +62,20 @@ class HotelController extends BaseController {
 
        public function joinHotel($id)
     {
-        if(Authority::getCurrentUser()->hasRole('member')){
+        if(Authority::can('join','Hotel')){
+            $user = User::find(Auth::id());
+            $hotel = hotel::find($id);
+            $user->requestHotels()->attach($hotel);
+            return Redirect::to('')->with('success', 'You have successfully request to join hotel');
+        }
+       /* if(Authority::getCurrentUser()->hasRole('member')){
         $user = User::find(Auth::id());
         $hotel = hotel::find($id);
         $user->hotels()->attach($hotel);
         $user->roles()->detach(2);
         $user->roles()->attach(4);
         return Redirect::to('myhotel/'.$id)->with('success', 'You have successfully join hotel');
-    }
+    }*/
     }
 
 }
